Moved IndexMap2.h into place as IndexMap.h, now that it works (and is fasters)
